+++ tom_d4rt_generator/lib/testing.dart
@@ -26,5 +26,6 @@ library;
 export 'src/testing/d4rt_test_result.dart';
 export 'src/testing/d4rt_tester.dart';
 export 'src/testing/example_freshness.dart';
+export 'src/testing/fixture_resolution_parity.dart';
 export 'src/testing/package_resolution.dart';
 export 'src/testing/issue_test_helper.dart';